ROCM-1254 - move cooperative_groups::reduce() to its own header
