+++ projects/clr/hipamd/include/hip/amd_detail/amd_hip_cooperative_groups.h
@@ -21,10 +21,6 @@
 #include <hip/amd_detail/hip_cooperative_groups_helper.h>
 #endif
 
-#if !defined(__HIPCC_RTC__)
-#include <type_traits>
-#endif
-
 namespace cooperative_groups {
 
 /** \brief The base type of all cooperative group types.
@@ -1335,76 +1331,6 @@ struct bit_or {
     return lhs | rhs;
   }
 };
-
-namespace impl {
-  template <typename T, typename U>
-  using is_param_type_same = __hip_internal::is_same<typename __hip_internal::remove_cvref<T>,
-                                                     typename __hip_internal::remove_cvref<U>>;
-
-  // we can call reduce() only the block tiles that have a compile-time size
-  template <class TyGroup>
-  struct isTiledGroup : __hip_internal::false_type {
-  };
-
-  template <unsigned int N, class ParentCGTy>
-  struct isTiledGroup<cooperative_groups::thread_block_tile<N, ParentCGTy>>
-    : __hip_internal::integral_constant<bool,
-          (N == 1  || N == 2  || N == 4  || N == 8 ||
-           N == 16 || N == 32 || N == 64)> {
-  };
-
-  template <class TyGroup>
-  struct isCoalescedGroup : __hip_internal::false_type {
-  };
-
-  template <>
-  struct isCoalescedGroup<cooperative_groups::coalesced_group> : __hip_internal::true_type {
-  };
-}
-
-template <typename TyGroup, typename TyVal, typename TyFn>
-__CG_QUALIFIER__ auto reduce(const TyGroup& group, TyVal&& val, TyFn&& op) -> decltype(op(val, val)) {
-  using Op = typename __hip_internal::remove_cvref<TyFn>::type;
-  using Val = typename __hip_internal::remove_cvref<TyVal>::type;
-  static_assert(impl::is_param_type_same<Val, decltype(op(val, val))>::value, "Operator input and output types differ");
-  static_assert(__hip_internal::is_trivially_copyable<Val>::value, "val must be trivially copyable");
-  // this is due to a current limitation on reduce_op_sync()
-  static_assert(sizeof(Val) <= 8, "Can only reduce values of up to 8 bytes");
-
-  if constexpr (!impl::isTiledGroup<TyGroup>::value && !impl::isCoalescedGroup<TyGroup>::value) {
-    static_assert(__hip_internal::is_void<TyGroup>::value, "This group does not exclusively represent a tile");
-  }
-
-  // we cannot simply use the __activemask() here, because more than one tile could have active
-  // threads at a time
-  unsigned long long mask = ~0ull;
-
-  if constexpr (!__hip_internal::is_same<TyGroup, cooperative_groups::coalesced_group>::value) {
-    mask >>= (64 - group.num_threads());
-    mask <<= (((threadIdx.x % warpSize) / group.num_threads()) * group.num_threads());
-  }
-
-  // for coalesced_groups, the mask is simply the activemask
-  // for tiled groups, it is legal for some threads in a tile to not participate so we also
-  // need to apply the active mask
-  mask &= __activemask();
-
-  if constexpr (__hip_internal::is_same<Op, cooperative_groups::plus<Val>>::value) {
-    return __reduce_add_sync(mask, val);
-  } else if constexpr (__hip_internal::is_same<Op, cooperative_groups::less<Val>>::value) {
-    return __reduce_min_sync(mask, val);
-  } else if constexpr (__hip_internal::is_same<Op, cooperative_groups::greater<Val>>::value) {
-    return __reduce_max_sync(mask, val);
-  } else if constexpr (__hip_internal::is_same<Op, cooperative_groups::bit_and<Val>>::value) {
-    return __reduce_and_sync(mask, val);
-  } else if constexpr (__hip_internal::is_same<Op, cooperative_groups::bit_or<Val>>::value) {
-    return __reduce_or_sync(mask, val);
-  } else if constexpr (__hip_internal::is_same<Op, cooperative_groups::bit_xor<Val>>::value) {
-    return __reduce_xor_sync(mask, val);
-  } else {
-    return __reduce_op_sync(mask, val, op, nullptr);
-  }
-}
 #endif
 }  // namespace cooperative_groups
