+++ projects/hip-tests/catch/unit/cooperativeGrps/cooperative_groups_common.hh
@@ -8,6 +8,7 @@
 
 #include <hip_test_common.hh>
 #include <hip/hip_cooperative_groups.h>
+#include <hip/cooperative_groups/hip_reduce.h>
 #include <cmd_options.hh>
 #include "../math/math_common.hh"
 namespace {
